+++ source/tools/maketech.cpp
@@ -279,7 +279,7 @@ main (int argc, char *argv[] )
 void *       loadpcx2(char *       filestring)
 {      
   void         *p = NULL;
-  byte         b; 
+  int b; 
 
 
    initgraphics (640, 480, 8);
